requirements.txt
# ACMGS Dependencies
# Install with: pip install -r requirements.txt

# ── Core Data Science ──
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ── Machine Learning ──
scikit-learn>=1.3.0
xgboost>=2.0.0

# ── Deep Learning (LSTM Autoencoder) ──
torch>=2.0.0

# ── Evolutionary Optimization ──
deap>=1.4.0

# ── API Layer ──
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# ── Dashboard ──
streamlit>=1.28.0
plotly>=5.18.0

# ── Visualization ──
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Utilities ──
python-dotenv>=1.0.0
tqdm>=4.65.0
pyserial>=3.5.0  # Arduino serial communication
aiohttp>=3.8.0  # Async HTTP for ESP32 client
websockets>=11.0.0  # WebSocket for real-time streaming
aiofiles>=23.0.0  # Async file operations
